backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Data validation
pydantic==2.10.6
pydantic-settings==2.7.1

# File upload support
python-multipart==0.0.20

# Environment variables
python-dotenv==1.0.1

# LLM (OpenAI-compatible SDK, works with 阿里百炼/Qwen etc.)
openai==1.59.9
rank-bm25==0.2.2

# Document parsing
pymupdf==1.25.3
python-docx==1.1.2
